requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
accelerate>=0.33.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
attrs==25.4.0
audioread==3.1.0
babel==2.17.0
bitsandbytes>0.37.0; platform_machine != "arm64" and platform_system != "Darwin"
boto3==1.42.16
botocore==1.42.16
brotli==1.2.0
cached_path
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click
contourpy==1.3.2
csvw==3.7.0
cycler==0.12.1
datasets
decorator==5.2.1
dill==0.4.0
dlinfo==2.0.0
docopt==0.6.2
einops==0.8.1
einx==0.3.0
ema-pytorch==0.7.3
encodec==0.1.1
espeakng==1.0.2
espeakng-loader==0.2.4
espeak_phonemizer==1.3.1
fastapi==0.127.0
ffmpy==1.0.0
filelock==3.20.1
fonttools==4.61.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.28.1
google-auth==2.45.0
google-cloud-core==2.5.0
google-cloud-storage==3.7.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
gradio==5.38.0
gradio-client==1.11.0
groovy==0.1.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
hydra-core>=1.3.0
idna==3.11
isodate==0.7.2
jieba
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langid==1.1.6
language-tags==1.2.0
lazy_loader==0.4
librosa
llvmlite==0.42.0
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe
matplotlib
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.18
networkx==3.1
num2words==0.5.13
numba==0.59.0
numpy==1.26.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
onnx==1.16.0
onnxruntime
onnxruntime-gpu
orjson==3.11.5
packaging==25.0
pandas==2.3.3
phonemizer==3.3.0
pillow==11.3.0
platformdirs==4.5.1
pooch==1.8.2
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.2
psutil==7.2.0
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic<=2.10.6
pydantic_core==2.27.2
pydub
py-espeak-ng==0.1.8
Pygments==2.19.2
pyparsing==3.3.1
pypinyin
pypinyin-dict
python-dateutil==2.9.0.post0
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
rdflib==7.5.0
referencing==0.37.0
regex
requests==2.32.5
rfc3986==1.5.0
rich==13.9.4
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
safehttpx==0.1.7
safetensors
scikit-learn==1.7.1
scipy==1.15.3
segments==2.3.0
semantic-version==2.10.0
sentry-sdk==2.48.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
soundfile
soxr==1.0.0
starlette==0.50.0
sympy==1.14.0
termcolor==3.2.0
threadpoolctl==3.6.0
tokenizers==0.22.1
tomli
tomlkit==0.13.3
torch==2.3.1
torchaudio==2.3.1
torchdiffeq==0.2.4
tqdm>=4.65.0
transformers
transformers-stream-generator
triton==2.3.1
typer==0.16.0
typing_extensions==4.12.2
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.2
uroman
uvicorn==0.40.0
vocos
x-transformers>=1.31.14
xxhash==3.6.0
yarl==1.22.0
zhconv